Move the Root config options to __init__.py
